requirements.txt
opencv-python
mediapipe
numpy
groq
python-dotenv
